+++ internal/attachments/templates/frontend/htmx/package.json.gotmpl
@@ -5,12 +5,12 @@
   "license": "MIT",
   "browserslist": "> 0.5%, last 2 versions, not dead",
   "source": [
-    "./assets/scripts.js",
-    "./assets/styles.css"
+    "./web/src/js/scripts.js",
+    "./web/src/scss/styles.scss"
   ],
   "targets": {
     "default": {
-      "distDir": "./static"
+      "distDir": "./web/static"
     }
   },
   "scripts": {
@@ -19,21 +19,18 @@
     "build": "parcel build"
   },
   "dependencies": {
+    {{ if or .CSSFramework "daisyui" "flowbite" "tailwindcss" }}
+    "@tailwindcss/typography": "latest",
+    "tailwindcss": "latest",
+    {{ end }}
     {{ if eq .CSSFramework "bootstrap" }}
     "bootstrap": "latest",
     {{ else if eq .CSSFramework "bulma" }}
     "bulma": "latest",
     {{ else if eq .CSSFramework "daisyui" }}
-    "@tailwindcss/typography": "latest",
-    "tailwindcss": "latest",
     "daisyui": "latest",
     {{ else if eq .CSSFramework "flowbite" }}
-    "@tailwindcss/typography": "latest",
-    "tailwindcss": "latest",
     "flowbite": "latest",
-    {{ else if eq .CSSFramework "tailwindcss" }}
-    "@tailwindcss/typography": "latest",
-    "tailwindcss": "latest",
     {{ else if eq .CSSFramework "unocss" }}
     "unocss": "latest",
     {{ end }}
@@ -42,23 +39,12 @@
   },
   "devDependencies": {
     {{ if eq .CSSFramework "bootstrap" }}
-    "@parcel/transformer-sass": "latest",
     "@popperjs/core": "latest",
-    {{ else if eq .CSSFramework "bulma" }}
-    "@parcel/transformer-sass": "latest",
-    {{ else if eq .CSSFramework "daisyui" }}
-    "postcss": "latest",
-    {{ else if eq .CSSFramework "flowbite" }}
-    "postcss": "latest",
-    {{ else if eq .CSSFramework "tailwindcss" }}
-    "postcss": "latest",
     {{ else if eq .CSSFramework "unocss" }}
     "@unocss/cli": "latest",
     "@unocss/postcss": "latest",
-    "postcss": "latest",
-    {{ else }}
-    "postcss": "latest",
     {{ end }}
+    "postcss": "latest",
     "parcel": "latest"
   }
 }
