+++ internal/attachments/templates/misc/air.toml.gotmpl
@@ -9,8 +9,8 @@ tmp_dir = "tmp"
     cmd = "{{ if .IsUseTempl }}templ generate && {{ end }}{{ if .IsUseBun }}bun{{ else }}npm{{ end }} run build && go build -o ./tmp/main ."
     post_cmd = ["echo 'Stopped Air live-reloading tool. See you next time!'"]
     bin = "tmp/main"
-    include_ext = ["go", "tpl", "tmpl", "gotmpl", "templ", "html"]
-    exclude_dir = ["assets", "bin", "dist", "tmp", "tests", "vendor", "node_modules"]
+    include_ext = ["go", "templ", "html", "js", "ts", "tsx", "vue", "svelte", "css", "scss"]
+    exclude_dir = ["bin", "dist", "dist-ssr", "tmp", "tests", "vendor", "node_modules"]
     include_dir = []
     include_file = []
     exclude_file = []
