+++ internal/forms/create.go
@@ -1,20 +1,66 @@
 package forms
 
 import (
+	"fmt"
+
 	"github.com/charmbracelet/huh"
 	"github.com/gowebly/gowebly/internal/fields"
 	"github.com/gowebly/gowebly/internal/injectors"
+	"github.com/gowebly/gowebly/internal/messages"
 )
 
-// CreateWelcomeForm runs the welcome form.
-func CreateWelcomeForm(di *injectors.Injector) error {
+// RunCreateForm runs the create form.
+func RunCreateForm(di *injectors.Injector) error {
+	// Run welcome note form.
+	if err := welcomeForm(); err != nil {
+		return fmt.Errorf(messages.ErrorFormNotRun, "welcome", "create", err)
+	}
+
+	// Run project settings form.
+	if err := projectSettingsForm(di); err != nil {
+		return fmt.Errorf(messages.ErrorFormNotRun, "project settings", "create", err)
+	}
+
+	// Run Go framework form.
+	if err := goFrameworkForm(di); err != nil {
+		return fmt.Errorf(messages.ErrorFormNotRun, "go framework", "create", err)
+	}
+
+	// Run HTMX form.
+	if err := htmxForm(di); err != nil {
+		return fmt.Errorf(messages.ErrorFormNotRun, "htmx", "create", err)
+	}
+
+	// Check, if HTMX is used.
+	if di.Config.Frontend.IsUseHTMX {
+		// If yes, run Templ form.
+		if err := templForm(di); err != nil {
+			return fmt.Errorf(messages.ErrorFormNotRun, "templ", "create", err)
+		}
+	} else {
+		// If not, run reactive library form.
+		if err := reactiveLibraryForm(di); err != nil {
+			return fmt.Errorf(messages.ErrorFormNotRun, "reactive library", "create", err)
+		}
+	}
+
+	// Run CSS framework form.
+	if err := cssFrameworkForm(di); err != nil {
+		return fmt.Errorf(messages.ErrorFormNotRun, "css framework", "create", err)
+	}
+
+	return nil
+}
+
+// welcomeForm runs the welcome form.
+func welcomeForm() error {
 	return huh.NewForm(
 		huh.NewGroup(fields.WelcomeNote()), // add welcome note
 	).Run()
 }
 
-// CreateProjectSettingsForm runs the project settings form.
-func CreateProjectSettingsForm(di *injectors.Injector) error {
+// projectSettingsForm runs the project settings form.
+func projectSettingsForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.GoModuleNameInput(di)), // input Go module name in go.mod
 		huh.NewGroup(fields.IsUseAirConfirm(di)),   // confirm if Air is used
@@ -23,36 +69,36 @@ func CreateProjectSettingsForm(di *injectors.Injector) error {
 	).Run()
 }
 
-// CreateGoFrameworkForm runs the Go framework form.
-func CreateGoFrameworkForm(di *injectors.Injector) error {
+// goFrameworkForm runs the Go framework form.
+func goFrameworkForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.GoFrameworkSelect(di)), // select Go framework
 	).Run()
 }
 
-// CreateHTMXForm runs the HTMX form.
-func CreateHTMXForm(di *injectors.Injector) error {
+// htmxForm runs the HTMX form.
+func htmxForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.IsUseHTMXConfirm(di)), // confirm if htmx is used
 	).Run()
 }
 
-// CreateTemplForm runs the Templ form.
-func CreateTemplForm(di *injectors.Injector) error {
+// templForm runs the Templ form.
+func templForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.IsUseTempleConfirm(di)), // confirm if Templ is used
 	).Run()
 }
 
-// CreateReactiveLibraryForm runs the reactive library form.
-func CreateReactiveLibraryForm(di *injectors.Injector) error {
+// reactiveLibraryForm runs the reactive library form.
+func reactiveLibraryForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.ReactiveLibrarySelect(di)), // select reactive library
 	).Run()
 }
 
-// CreateCSSFrameworkForm runs the CSS framework form.
-func CreateCSSFrameworkForm(di *injectors.Injector) error {
+// cssFrameworkForm runs the CSS framework form.
+func cssFrameworkForm(di *injectors.Injector) error {
 	return huh.NewForm(
 		huh.NewGroup(fields.CSSFrameworkSelect(di)), // select CSS framework
 	).Run()
